Refactor: avatar label 컴포넌트 공통화 및 공통 컴포넌트로 변경
